Split up tests for parallel execution
